Accelerometer driving 8x8 LED and a LCD screen as it moves
Accelerometer is on a bread board. As user moves the accelerometer in the free space the 8x8 LED matrix will show the current position. Upwards position shows no LED and when user tilts the accelerometer the LED shows the tilt. Also current tilt position is shown as x,y,z degrees in LCD screen. All of the logic and configrations of the peripherals done in TIVA Board TM123C.